Realy fix timetable order bug

+++ WebApp/Models/StudentStatDTO.cs
@@ -12,16 +12,16 @@ namespace NFCSystem.Models
         public string LectureID { get; set; }
         public string LectureName { get; set; }
 
-        public int AttendanceLab { get; set; }
+        public double AttendanceLab { get; set; }
         public int TotalLab { get; set; }
         public int VisitedLab { get; set; }
     
 
-        public int AttendanceTheory { get; set; }
+        public double AttendanceTheory { get; set; }
         public int TotalTheory { get; set; }
         public int VisitedTheory { get; set; }
 
-        public int AttendancePractice { get; set; }
+        public double AttendancePractice { get; set; }
         public int TotalPractice {get;set;}
         public int VisitedPractice {get;set;}
     }
